extract source_statuses
